moved check_subnet to tests folder
