+++ scripts/check_environment.py
@@ -4,7 +4,7 @@
 configure the Bluesky framework
 """
 
-from aps_8id_bs_instrument.utils.check_subnet import warn_if_not_aps_controls_subnet
+from aps_8id_bs_instrument.tests.check_subnet import warn_if_not_aps_controls_subnet
 from aps_8id_bs_instrument.utils.metadata import metadata
 
 warn_if_not_aps_controls_subnet()
